refactor: Add constants file
